Refactor code #1
Bugs:
-router.param not working
- pagination displays page 2 with data even if total pages is only 1
