--- routes/utils/auth.js
@@ -1,99 +1,0 @@
-const User = require("../../models/User");
-
-const router = require("express").Router();
-const CryptoJS = require("crypto-js");
-const jwt = require("jsonwebtoken");
-
-// Register
-router.post("/register", async (req, res) => {
-  const newUser = new User({
-    username: req.body.username,
-    password: CryptoJS.AES.encrypt(
-      req.body.password,
-      process.env.PASSPHRASE
-    ).toString(),
-    email: req.body.email,
-  });
-
-  try {
-    const savedUser = await newUser.save();
-    return res.status(201).json({
-      status: 201,
-      message: `Created user. User ID: ${savedUser._id}`,
-    });
-  } catch (error) {
-    console.log(error);
-    if (
-      error.name === "MongoServerError" &&
-      error.code === 11000 &&
-      error.keyValue.hasOwnProperty("email")
-    ) {
-      return res.status(400).json({
-        status: 400,
-        message: `Email already exists.`,
-      });
-    } else if (error.name === "MongoServerError" && error.code === 11000) {
-      return res.status(400).json({
-        status: 400,
-        message: `Username already exists.`,
-      });
-    } else if (error.name === "ValidationError") {
-      return res.status(400).json({
-        status: 400,
-        message: error.message,
-      });
-    } else return res.status(500).json(error);
-  }
-});
-
-// Login
-router.post("/login", async (req, res) => {
-  try {
-    const user = await User.findOne({ username: req.body.username });
-    if (!user) {
-      return res.status(401).json({
-        status: 401,
-        message: "Invalid username/password.",
-      });
-    }
-
-    const hashedPassword = CryptoJS.AES.decrypt(
-      user.password,
-      process.env.PASSPHRASE
-    );
-    const decryptedPassword = hashedPassword.toString(CryptoJS.enc.Utf8);
-
-    if (decryptedPassword !== req.body.password) {
-      return res.status(401).json({
-        status: 401,
-        message: "Invalid username/password.",
-      });
-    }
-
-    const accessToken = jwt.sign(
-      {
-        id: user._id,
-        isAdmin: user.isAdmin,
-      },
-      process.env.JWT_SEC,
-      { expiresIn: "3d" }
-    );
-
-    const { _id, username, isAdmin, ...hide } = user._doc;
-
-    return res.status(200).json({
-      status: 200,
-      message: "Login successful!",
-      result: {
-        _id,
-        username,
-        isAdmin,
-        accessToken
-      }
-    });
-  } catch (error) {
-    return res.status(500).json(error);
-  }
-});
-
-module.exports = router;
